requirements.txt
asgiref==3.10.0
asttokens==3.0.0
attrs==25.4.0
Authlib==1.6.5
bandit==1.7.6
black==23.12.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
coverage==7.11.0
cryptography==46.0.3
decorator==5.2.1
Django==5.2.7
django-appconf==1.1.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==25.2
djangorestframework==3.14.0
dparse==0.6.4
drf-spectacular==0.27.2
executing==2.2.1
factory-boy==3.3.0
Faker==37.11.0
flake8==7.0.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
ipython==8.19.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.0.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mypy==1.7.1
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.2.3
psycopg-binary==3.2.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.11.1
pycparser==2.23
pydantic==1.10.24
pyflakes==3.2.0
Pygments==2.19.2
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
safety==3.0.1
safety-schemas==0.0.5
setuptools==80.9.0
shellingham==1.5.4
smmap==5.0.2
sqlparse==0.5.3
stack-data==0.6.3
stevedore==5.5.0
traitlets==5.14.3
typer==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
